+++ python/cmtj.cpp
@@ -8,6 +8,8 @@
 #include "../core/cvector.hpp"
 #include "../core/drivers.hpp"
 #include "../core/junction.hpp"
+#include "../core/noise.hpp"
+#include "../core/llgb.hpp"
 #include <stdio.h>
 #include <vector>
 
@@ -15,6 +17,8 @@ using namespace pybind11::literals;
 namespace py = pybind11;
 
 using DJunction = Junction<double>;
+using DParallelStack = ParallelStack<double>;
+using DSeriesStack = SeriesStack<double>;
 using DLayer = Layer<double>;
 using DVector = CVector<double>;
 using DScalarDriver = ScalarDriver<double>;
@@ -23,6 +27,8 @@ using DNullDriver = NullDriver<double>;
 using DStack = Stack<double>;
 using DLayerMatrix = std::vector<std::vector<DLayer>>;
 using DVectorMatrix = std::vector<std::vector<DVector>>;
+using DLLGBLayer = LLGBLayer<double>;
+using DLLGBJunction = LLGBJunction<double>;
 
 #define USING_PY true
 PYBIND11_MODULE(cmtj, m)
@@ -53,8 +59,8 @@ PYBIND11_MODULE(cmtj, m)
         .def(py::self * double())
         .def("__getitem__", [](const DVector& v, const int key) { return v[key]; })
         .def("__len__", [](const DVector& v) { return 3; })
-        .def("__str__", &DVector::toString)
-        .def("__repr__", &DVector::toString);
+        .def("__str__", py::overload_cast<>(&DVector::toString))
+        .def("__repr__", py::overload_cast<>(&DVector::toString));
 
     py::implicitly_convertible<std::list<double>, DVector>();
     py::implicitly_convertible<std::vector<double>, DVector>();
@@ -63,6 +69,8 @@ PYBIND11_MODULE(cmtj, m)
         .value("xaxis", xaxis)
         .value("yaxis", yaxis)
         .value("zaxis", zaxis)
+        .value("all", all)
+        .value("none", none)
         .export_values();
 
     py::enum_<Reference>(m, "Reference")
@@ -205,8 +213,13 @@ PYBIND11_MODULE(cmtj, m)
         .def_readonly("damping", &DLayer::damping)
         .def_readonly("cellSurface", &DLayer::cellSurface)
         .def_readonly("demagTensor", &DLayer::demagTensor)
+        // noise
+        .def("setAlphaNoise", &DLayer::setAlphaNoise)
+        .def("setOneFNoise", &DLayer::setOneFNoise)
         // getters
-        .def("getId", &DLayer::getId);
+        .def("getId", &DLayer::getId)
+        .def("getOneFVector", &DLayer::getOneFVector)
+        .def("createBufferedAlphaNoise", &DLayer::createBufferedAlphaNoise);
 
     py::class_<DJunction>(m, "Junction")
         .def(py::init<std::vector<DLayer>>(),
@@ -277,51 +290,45 @@ PYBIND11_MODULE(cmtj, m)
 
     // stack module
     py::module stack_module = m.def_submodule("stack", "A stack submodule for joining MTJ junctions");
-    py::class_<SeriesStack<double>>(stack_module, "SeriesStack")
+    py::class_<DSeriesStack>(stack_module, "SeriesStack")
         .def(py::init<std::vector<DJunction>,
             std::string,
             std::string>(),
             "junctionList"_a,
             "topId_a"_a = "free",
             "bottomId"_a = "bottom")
-        .def("runSimulation", &SeriesStack<double>::runSimulation,
+        .def("runSimulation", &DSeriesStack::runSimulation,
             "totalTime"_a,
             "timeStep"_a = 1e-13,
             "writeFrequency"_a = 1e-11)
-        .def("runSimulationNonDelayed", &SeriesStack<double>::runSimulationNonDelayed,
-            "totalTime"_a,
-            "timeStep"_a = 1e-13,
-            "writeFrequency"_a = 1e-11)
-        .def("setMagnetisation", &SeriesStack<double>::setMagnetisation, "junction"_a, "layerId"_a, "mag"_a)
-        .def("getMagnetisation", &SeriesStack<double>::getMagnetisation, "junction"_a, "layerId"_a)
-        .def("setCoupledCurrentDriver", &SeriesStack<double>::setCoupledCurrentDriver, "driver"_a)
-        .def("setExternalFieldDriver", &SeriesStack<double>::setExternalFieldDriver, "driver"_a)
-        .def("setCouplingStrength", &SeriesStack<double>::setCouplingStrength, "coupling"_a)
+        .def("setMagnetisation", &DSeriesStack::setMagnetisation, "junction"_a, "layerId"_a, "mag"_a)
+        .def("getMagnetisation", &DSeriesStack::getMagnetisation, "junction"_a, "layerId"_a)
+        .def("setCoupledCurrentDriver", &DSeriesStack::setCoupledCurrentDriver, "driver"_a)
+        .def("setExternalFieldDriver", &DSeriesStack::setExternalFieldDriver, "driver"_a)
+        .def("setCouplingStrength", &DSeriesStack::setCouplingStrength, "coupling"_a)
+        .def("setDelayed", &DSeriesStack::setDelayed, "delayed"_a)
         // logging
-        .def("clearLogs", &SeriesStack<double>::clearLogs)
-        .def("getLog", py::overload_cast<unsigned int>(&SeriesStack<double>::getLog))
-        .def("getLog", py::overload_cast<>(&SeriesStack<double>::getLog));
+        .def("clearLogs", &DSeriesStack::clearLogs)
+        .def("getLog", py::overload_cast<unsigned int>(&DSeriesStack::getLog))
+        .def("getLog", py::overload_cast<>(&DSeriesStack::getLog));
 
-    py::class_<ParallelStack<double>>(stack_module, "ParallelStack")
+    py::class_<DParallelStack>(stack_module, "ParallelStack")
         .def(py::init<std::vector<DJunction>,
             std::string,
             std::string>(),
             "junctionList"_a,
             "topId_a"_a = "free",
             "bottomId"_a = "bottom")
-        .def("runSimulation", &ParallelStack<double>::runSimulation,
+        .def("runSimulation", &DParallelStack::runSimulation,
             "totalTime"_a,
             "timeStep"_a = 1e-13,
             "writeFrequency"_a = 1e-11)
-        .def("runSimulationNonDelayed", &ParallelStack<double>::runSimulationNonDelayed,
-            "totalTime"_a,
-            "timeStep"_a = 1e-13,
-            "writeFrequency"_a = 1e-11)
-        .def("setMagnetisation", &ParallelStack<double>::setMagnetisation, "junction"_a, "layerId"_a, "mag"_a)
-        .def("getMagnetisation", &ParallelStack<double>::getMagnetisation, "junction"_a, "layerId"_a)
-        .def("setCoupledCurrentDriver", &ParallelStack<double>::setCoupledCurrentDriver, "driver"_a)
-        .def("setExternalFieldDriver", &ParallelStack<double>::setExternalFieldDriver, "driver"_a)
-        .def("setCouplingStrength", &ParallelStack<double>::setCouplingStrength, "coupling"_a)
+        .def("setMagnetisation", &DParallelStack::setMagnetisation, "junction"_a, "layerId"_a, "mag"_a)
+        .def("getMagnetisation", &DParallelStack::getMagnetisation, "junction"_a, "layerId"_a)
+        .def("setCoupledCurrentDriver", &DParallelStack::setCoupledCurrentDriver, "driver"_a)
+        .def("setExternalFieldDriver", &DParallelStack::setExternalFieldDriver, "driver"_a)
+        .def("setCouplingStrength", &DParallelStack::setCouplingStrength, "coupling"_a)
+        .def("setDelayed", &DParallelStack::setDelayed, "delayed"_a)
         // logging
         .def("clearLogs", &ParallelStack<double>::clearLogs)
         .def("getLog", py::overload_cast<unsigned int>(&ParallelStack<double>::getLog))
@@ -342,4 +349,72 @@ PYBIND11_MODULE(cmtj, m)
         .def("setLayerAnisotropy", &Reservoir::setLayerAnisotropy)
         .def("setLayerExternalField", &Reservoir::setLayerExternalField)
         .def("getMagnetisation", &Reservoir::getMagnetisation);
+
+    // generator module
+    py::module generator_module = m.def_submodule("noise", "Submodule with noise generation functions");
+    py::class_<BufferedAlphaNoise<double>>(generator_module, "BufferedAlphaNoise")
+        .def(py::init<unsigned int, double, double, double>(),
+            "bufferSize"_a,
+            "alpha"_a,
+            "std"_a,
+            "scale"_a)
+        .def("fillBuffer", &BufferedAlphaNoise<double>::fillBuffer)
+        .def("tick", &BufferedAlphaNoise<double>::tick);
+    py::class_<VectorAlphaNoise<double>>(generator_module, "VectorAlphaNoise")
+        .def(py::init<unsigned int, double, double, double, Axis>(),
+            "bufferSize"_a,
+            "alpha"_a,
+            "std"_a,
+            "scale"_a,
+            "axis"_a = Axis::all)
+        .def("tickVector", &VectorAlphaNoise<double>::tickVector)
+        .def("tick", &VectorAlphaNoise<double>::tick)
+        .def("getPrevSample", &VectorAlphaNoise<double>::getPrevSample)
+        .def("getScale", &VectorAlphaNoise<double>::getScale);
+
+
+    // LLGB module
+    py::module llgb_module = m.def_submodule("llgb", "A submodule for LLGB junctions");
+    py::class_<DLLGBLayer>(llgb_module, "LLGBLayer")
+        .def(py::init<const std::string&,
+            DVector,
+            DVector,
+            double,
+            double,
+            double,
+            const std::vector<DVector>&,
+            double,
+            double,
+            double,
+            double >(),
+            "id"_a,
+            "mag"_a,
+            "anis"_a,
+            "Ms"_a,
+            "thickness"_a,
+            "cellSurface"_a,
+            "demagTensor"_a,
+            "damping"_a,
+            "Tc"_a,
+            "susceptibility"_a,
+            "me"_a)
+        // setters
+        .def("setTemperatureDriver", &DLLGBLayer::setTemperatureDriver)
+        .def("setExternalFieldDriver", &DLLGBLayer::setExternalFieldDriver)
+        .def("setAnisotropyDriver", &DLLGBLayer::setAnisotropyDriver);
+
+    py::class_<DLLGBJunction>(llgb_module, "LLGBJunction")
+        .def(py::init<std::vector<DLLGBLayer>>(),
+            "layers"_a)
+        .def("runSimulation", &DLLGBJunction::runSimulation,
+            "totalTime"_a,
+            "timeStep"_a = 1e-13,
+            "writeFrequency"_a = 1e-11,
+            "log"_a = false,
+            "solverMode"_a = HEUN)
+        .def("setLayerTemperatureDriver", &DLLGBJunction::setLayerTemperatureDriver)
+        .def("setLayerExternalFieldDriver", &DLLGBJunction::setLayerExternalFieldDriver)
+        .def("saveLogs", &DLLGBJunction::saveLogs)
+        .def("getLog", &DLLGBJunction::getLog)
+        .def("clearLog", &DLLGBJunction::clearLog);
 }